+++ contracts/test/TestQuery.sol
@@ -8,14 +8,13 @@ using Cursors for Cur;
 using Writers for Writer;
 
 contract TestQuery is QueryBase {
-    bytes4 private immutable Value = Keys.local(1);
-    bytes32 private constant ValueName = bytes32("value");
-    string private constant ValueSchema = "{ uint value }";
+    bytes4 private constant Value = Keys.Local;
+    string private constant INPUT = "{ uint value }";
 
     bytes32 private immutable descriptor;
 
     constructor() {
-        (, descriptor) = query("incrementQuery", schema(Value, ValueSchema, ValueName), Value, 0);
+        (, descriptor) = query("incrementQuery", localSchema(INPUT), Value, 0);
     }
 
     function incrementQuery(bytes calldata request) external view returns (bytes memory out) {
@@ -25,6 +24,30 @@ contract TestQuery is QueryBase {
         while (input.i < input.len) {
             uint foo = uint(input.unpack32(Value));
             output.appendBlock32(Value, bytes32(foo + 1), 32);
+        }
+
+        input.complete();
+        out = output.finish();
+    }
+}
+
+contract TestKeyedLocalQuery is QueryBase {
+    bytes4 private immutable Value = Keys.local(2);
+    string private constant INPUT = "{ uint value }";
+
+    bytes32 private immutable descriptor;
+
+    constructor() {
+        (, descriptor) = query("keyedLocalQuery", localSchema(2, INPUT), Value, 0);
+    }
+
+    function keyedLocalQuery(bytes calldata request) external view returns (bytes memory out) {
+        (Cur memory input, uint outputs) = openInput(request, descriptor);
+        Writer memory output = Writers.alloc32s(outputs);
+
+        while (input.i < input.len) {
+            uint foo = uint(input.unpack32(Value));
+            output.appendBlock32(Value, bytes32(foo + 2), 32);
         }
 
         input.complete();